// src/types.tsx
export type AlgoPostDataProps = {
  id?: string;
  lots?: string;
  position?: string;
  optionType?: string;
  expiry?: string;
  strikeCriteria?: string;
  atmStrike?: string;
  strikeAdjustMent?: string;
  lowerRange?: number;
  upperRange?: number;
  strikeType?: string;
  simpleMomentumType?: string;
  simpleMomentumValue?: number;
  trailSlType?: string;
  instrumentMove?: number;
  stopLossMove?: number;
};